refact: remove copied functino from developer_tool, remove buggy initialize repo
